Added current Javascript version of the flextile library plus tests.
The Javascript is currently the only version of the flextile library. Parsing and interpreting works but edge cases are not explored as well as I would like them to be.

flextile is pretty much useless without an runtime environment that defines the basic commands. An example enviroment (the one I use on a daily basis) can be found in the test folder.